backend/requirements.txt
fastapi
uvicorn
neo4j
openai
pandas
tabulate
matplotlib
networkx
python-multipart
langchain-neo4j
python-dotenv
